实现语言切换功能

+++ View-Spot-of-City/View-Spot-of-City.MapView/MapView.xaml.cs
@@ -34,9 +34,9 @@ namespace View_Spot_of_City.MapView
             mapControl.Position = new PointLatLng(Convert.ToDouble(Config.AppSettings["MAP_CENTER_LAT"]), Convert.ToDouble(Config.AppSettings["MAP_CENTER_LNG"]));
             mapControl.ShowCenter = false;
             mapControl.MouseWheelZoomType = MouseWheelZoomType.MousePositionWithoutCenter;
-            mapControl.MinZoom = Convert.ToInt32(Config.AppSettings["MAP_MIN_ZOOM"]);
-            mapControl.MaxZoom = Convert.ToInt32(Config.AppSettings["MAP_MAX_ZOOM"]);
-            mapControl.Zoom = Convert.ToInt32(Config.AppSettings["MAP_ZOOM"]);
+            mapControl.MinZoom = (int)Convert.ToDouble(Config.AppSettings["MAP_MIN_ZOOM"]);
+            mapControl.MaxZoom = (int)Convert.ToDouble(Config.AppSettings["MAP_MAX_ZOOM"]);
+            mapControl.Zoom = Convert.ToDouble(Config.AppSettings["MAP_ZOOM"]);
             mapControl.ShowCenter = false;
             mapControl.DragButton = MouseButton.Left;
             mapControl.Position = new PointLatLng(Convert.ToDouble(Config.AppSettings["MAP_CENTER_LAT"]), Convert.ToDouble(Config.AppSettings["MAP_CENTER_LNG"]));
